+++ icloudpd/base.py
@@ -128,7 +128,8 @@ CONTEXT_SETTINGS = dict(help_option_names=["-h", "--help"])
 )
 @click.option(
     "--folder-structure",
-    help="Folder structure (default: {:%Y/%m/%d})",
+    help="Folder structure (default: {:%Y/%m/%d}). "
+    "If set to 'none' all photos will just be placed into the download directory",
     metavar="<folder_structure>",
     default="{:%Y/%m/%d}",
 )
@@ -390,7 +391,10 @@ def main(
             created_date = photo.created
 
         try:
-            date_path = folder_structure.format(created_date)
+            if folder_structure.lower() == "none":
+                date_path = ""
+            else:
+                date_path = folder_structure.format(created_date)
         except ValueError:  # pragma: no cover
             # This error only seems to happen in Python 2
             logger.set_tqdm_description(