+++ CHANGELOG.md
@@ -1,5 +1,9 @@
+# Changelog
+
 ## Unreleased
 
+- feature: Folder structure can be set to 'none' instead of a date pattern,
+so all photos will be placed directly into the download directory.
 - fix: Empty directory structure being created #185
 
 ## 1.6.2 (2020-10-23)
@@ -13,5 +17,5 @@
 
 ## Earlier Versions
 
-Please refer to the commit history in GitHub: 
-https://github.com/icloud-photos-downloader/icloud_photos_downloader/commits/master
+Please refer to the commit history in GitHub:
+<https://github.com/icloud-photos-downloader/icloud_photos_downloader/commits/master>